clippy: Squish warnings and errors in gfx
warning: redundant closure (gfx/font.rs:415:18)

warning: useless conversion to the same type (gfx/font.rs:534:9)

warning: the following explicit lifetimes could be elided: 'a (gfx/font.rs:619:16)

error: this loop never actually loops (gfx/font_cache_thread.rs:112:9)

warning: this expression creates a reference which is immediately dereferenced by the compiler  (gfx/font_cache_thread.rs:229:51)

warning: redundant closure (gfx/font_cache_thread.rs:551:18)

3 instances of:
warning: casting integer literal to `f64` is unnecessary (gfx/platform/freetype/font_list.rs:271-273)

+++ components/gfx/platform/freetype/font_list.rs
@@ -268,9 +268,9 @@ fn font_weight_from_fontconfig_pattern(pattern: *mut FcPattern) -> Option<FontWe
 
     let mapping = [
         (0., 0.),
-        (FC_WEIGHT_REGULAR as f64, 400 as f64),
-        (FC_WEIGHT_BOLD as f64, 700 as f64),
-        (FC_WEIGHT_EXTRABLACK as f64, 1000 as f64),
+        (FC_WEIGHT_REGULAR as f64, 400_f64),
+        (FC_WEIGHT_BOLD as f64, 700_f64),
+        (FC_WEIGHT_EXTRABLACK as f64, 1000_f64),
     ];
 
     let mapped_weight = map_platform_values_to_style_values(&mapping, weight as f64);
